requirements.txt
fastapi
pydantic
python-jose
passlib
python-multipart
uvicorn
jose
python-dotenv
supabase
email-validator
bcrypt